Improve API access

+++ A32NX/html_ui/Pages/A32NX_Utils/NXApi.js
@@ -1,19 +1,18 @@
-NXApi.url = "https://api.flybywiresim.com";
-NXApi.telexNotEnabledError = "TELEX NOT ENABLED";
-NXApi.noRecipientError = "NO RECIPIENT SET";
-
 class NXApi {
     static getMetar(icao, source) {
         if (!icao) {
             throw ("No ICAO provided");
         }
 
+        console.log('GET METAR');
         return fetch(`${NXApi.url}/metar/${icao}?source=${source}`)
             .then((response) => {
                 if (!response.ok) {
+                    console.error('METAR RESPONSE ERROR');
                     throw (response);
                 }
 
+                console.log('METAR RESPONSE OK');
                 return response.json();
             });
     }
@@ -23,12 +22,15 @@ class NXApi {
             throw ("No ICAO provided");
         }
 
+        console.log('GET TAF');
         return fetch(`${NXApi.url}/taf/${icao}?source=${source}`)
             .then((response) => {
                 if (!response.ok) {
+                    console.error('TAF RESPONSE ERROR');
                     throw (response);
                 }
 
+                console.error('TAF RESPONSE OK');
                 return response.json();
             });
     }
@@ -38,12 +40,15 @@ class NXApi {
             throw ("No ICAO provided");
         }
 
+        console.log('GET ATIS');
         return fetch(`${NXApi.url}/atis/${icao}?source=${source}`)
             .then((response) => {
                 if (!response.ok) {
+                    console.error('ATIS RESPONSE ERROR');
                     throw (response);
                 }
 
+                console.log('ATIS RESPONSE OK');
                 return response.json();
             });
     }
@@ -51,21 +56,25 @@ class NXApi {
     static connectTelex(flightNo) {
         // TELEX disabled
         if (NXDataStore.get("CONFIG_TELEX_STATUS", "DISABLED") !== "ENABLED") {
-            throw (NXApi.telexNotEnabledError);
+            return Promise.reject(NXApi.disabledError);
         }
 
         const connectBody = NXApi.buildTelexBody(flightNo);
         const headers = { "Content-Type": "application/json" };
 
+        console.log("CONNECTING TO TELEX");
         return fetch(`${NXApi.url}/txcxn`, { method: "POST", body: JSON.stringify(connectBody), headers })
             .then((response) => {
                 if (!response.ok) {
+                    console.error("TELEX CONNECTION ERROR");
                     throw (response);
                 }
 
+                console.log("TELEX CONNECTION OK");
                 const data = response.json();
-                NXDataStore.set("TELEX_KEY", data["accessToken"]);
-                NXDataStore.set("TELEX_FLIGHT_NUMBER", data["flight"]);
+
+                NXDataStore.set("TELEX_KEY", data.accessToken);
+                NXDataStore.set("TELEX_FLIGHT_NUMBER", data.flight);
 
                 return data;
             });
@@ -74,28 +83,41 @@ class NXApi {
     static updateTelex() {
         // TELEX disabled
         if (NXDataStore.get("CONFIG_TELEX_STATUS", "DISABLED") !== "ENABLED") {
-            throw (NXApi.telexNotEnabledError);
+            return Promise.reject(NXApi.disabledError);
+        }
+
+        // No connection
+        if (!NXApi.hasTelexConnection()) {
+            return Promise.reject(NXApi.disconnectedError);
         }
 
         const updateBody = NXApi.buildTelexBody();
         const headers = {
             "Content-Type": "application/json",
-            Authorization: `Bearer ${NXDataStore.get("TELEX_KEY", "")}`
+            Authorization: NXApi.buildToken()
         };
 
+        console.log('UPDATING TELEX');
         return fetch(`${NXApi.url}/txcxn`, { method: "PUT", body: JSON.stringify(updateBody), headers })
             .then((response) => {
                 if (!response.ok) {
+                    console.error("TELEX UPDATE ERROR");
                     throw (response);
                 }
 
+                console.log("TELEX UPDATE OK");
                 return response.json();
             });
     }
 
     static disconnectTelex() {
+        // No connection
+        if (!NXApi.hasTelexConnection()) {
+            return Promise.reject(NXApi.disabledError);
+        }
+
         const headers = {
-            Authorization: `Bearer ${NXDataStore.get("TELEX_KEY", "")}`
+            Authorization: NXApi.buildToken()
         };
 
         return fetch(`${NXApi.url}/txcxn`, { method: "DELETE", headers })
@@ -104,22 +126,26 @@ class NXApi {
                     throw (response);
                 }
 
-                const data = response.json();
                 NXDataStore.delete("TELEX_KEY");
                 NXDataStore.delete("TELEX_FLIGHT_NUMBER");
 
-                return data;
+                return response.json();
             });
     }
 
     static getTelexMessages() {
         // TELEX disabled
         if (NXDataStore.get("CONFIG_TELEX_STATUS", "DISABLED") !== "ENABLED") {
-            throw (NXApi.telexNotEnabledError);
+            return Promise.reject(NXApi.disabledError);
+        }
+
+        // No connection
+        if (!NXApi.hasTelexConnection()) {
+            return Promise.reject(NXApi.disconnectedError);
         }
 
         const headers = {
-            Authorization: `Bearer ${NXDataStore.get("TELEX_KEY", "")}`
+            Authorization: NXApi.buildToken()
         };
 
         return fetch(`${NXApi.url}/txmsg`, { method: "GET", headers })
@@ -133,8 +159,19 @@ class NXApi {
     }
 
     static sendTelexMessage(recipient, message) {
+        // TELEX disabled
+        if (NXDataStore.get("CONFIG_TELEX_STATUS", "DISABLED") !== "ENABLED") {
+            return Promise.reject(NXApi.disabledError);
+        }
+
+        // No connection
+        if (!NXApi.hasTelexConnection()) {
+            return Promise.reject(NXApi.disconnectedError);
+        }
+
+        // No recipient
         if (!recipient) {
-            throw (NXApi.noRecipientError);
+            return Promise.reject(NXApi.noRecipientError);
         }
 
         const body = {
@@ -143,7 +180,7 @@ class NXApi {
         };
         const headers = {
             "Content-Type": "application/json",
-            Authorization: `Bearer ${NXDataStore.get("TELEX_KEY", "")}`
+            Authorization: NXApi.buildToken()
         };
 
         return fetch(`${NXApi.url}/txmsg`, {method: "POST", body: JSON.stringify(body), headers})
@@ -155,7 +192,18 @@ class NXApi {
     }
 
     static hasTelexConnection() {
-        return NXDataStore.get("TELEX_KEY", "") && NXDataStore.get("TELEX_FLIGHT_NUMBER", "");
+        const txKey = NXDataStore.get("TELEX_KEY", "");
+        console.log("KEY." + txKey);
+
+        const txFlight = NXDataStore.get("TELEX_FLIGHT_NUMBER", "");
+        console.log("FLT." + txFlight);
+
+        return txKey && txFlight;
+    }
+
+    static buildToken() {
+        const txKey = NXDataStore.get("TELEX_KEY", "");
+        return `Bearer ${txKey}`;
     }
 
     static buildTelexBody(flightNo) {
@@ -177,3 +225,8 @@ class NXApi {
         };
     }
 }
+
+NXApi.url = "https://fbw.stonelabs.io";
+NXApi.disabledError = "TELEX DISABLED";
+NXApi.disconnectedError = "TELEX DISCONNECTED";
+NXApi.noRecipientError = "NO RECIPIENT";
